+++ src/app/app-routing.module.ts
@@ -1,18 +1,13 @@
-import { VuelodComponent } from './vuelod/vuelod.component';
-import { TablaVDisponibleComponent } from './tabla-vdisponible/tabla-vdisponible.component';
 import { PrincipalComponent } from './principal/principal.component';
-import { MostrarTablaComponent } from './componentes/vuelo/mostrar-tabla/mostrar-tabla.component';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { TablaVDisponibleComponent } from './componentes/tabla-vdisponible/tabla-vdisponible.component';
+
 
 const routes: Routes = [
   {path:'', redirectTo:'/inicio',pathMatch:'full'},
   {path:'inicio', component: PrincipalComponent},
-  {path:'vuelo', component: MostrarTablaComponent},
   {path: 'disponible', component: TablaVDisponibleComponent},
-  {path: 'disponibles', component: VuelodComponent},
-
-
 
 ];
