+++ src/app/componentes/interfaces/ruta.ts
@@ -1,10 +1,12 @@
 export interface Ruta {
   nombre: string,
-  origen:string,
-  destino:string,
+  origen:any,
+  destino:any,
   distancia_viaje:string,
   tiempo_viaje:string,
   precio_base:string,
-  fecha_salida:string,
+  fecha_salida:any,
   descripcion:string
+  id_vuelo:string;
+  id_ruta:string;
 }
