+++ src/app/componentes/interfaces/vuelo.ts
@@ -1,0 +1,12 @@
+export interface Vuelo{
+  id_vuelo:string;
+  id_avion:string;
+  id_ruta:string;
+  origen:string;
+  destino:string;
+  precio_base:string;
+  distancia_viaje:string;
+  tiempo_viaje:string;
+  fecha_creacion:string;
+  descripcion:string;
+}
