+++ src/app/componentes/tabla-vdisponible/tabla-vdisponible.component.html
@@ -2,62 +2,54 @@
   <mat-toolbar style="background-color:rgb(241, 218, 87);">
     <span>Vuelos Disponibles</span>
     <div class="example-spacer"></div>
-    <!--button mat-raised-button color="primary" routerLink="/agregar">Agregar</button-->
   </mat-toolbar>
 
   <table mat-table class="full-width-table" matSort aria-label="Elements">
-    <!-- Id Column -->
     <ng-container matColumnDef="nombre">
       <th mat-header-cell *matHeaderCellDef mat-sort-header>Aerolínea</th>
       <td mat-cell *matCellDef="let row">{{row.nombre}}</td>
     </ng-container>
 
-    <!-- Name Column -->
     <ng-container matColumnDef="origen">
       <th mat-header-cell *matHeaderCellDef mat-sort-header>Origen</th>
       <td mat-cell *matCellDef="let row">{{row.origen}}</td>
     </ng-container>
 
-    <!-- Name Column -->
     <ng-container matColumnDef="destino">
       <th mat-header-cell *matHeaderCellDef mat-sort-header>Destino</th>
       <td mat-cell *matCellDef="let row">{{row.destino}}</td>
     </ng-container>
 
-    <!-- Name Column -->
     <ng-container matColumnDef="distancia_viaje">
       <th mat-header-cell *matHeaderCellDef mat-sort-header>Distancia</th>
       <td mat-cell *matCellDef="let row">{{row.distancia_viaje}}</td>
     </ng-container>
 
-    <!-- Name Column -->
     <ng-container matColumnDef="tiempo_viaje">
       <th mat-header-cell *matHeaderCellDef mat-sort-header>Tiempo Viaje</th>
       <td mat-cell *matCellDef="let row">{{row.tiempo_viaje}}</td>
     </ng-container>
 
-    <!-- Name Column -->
     <ng-container matColumnDef="precio_base">
       <th mat-header-cell *matHeaderCellDef mat-sort-header>Precio Base</th>
       <td mat-cell *matCellDef="let row">{{row.precio_base}}</td>
     </ng-container>
 
-    <!-- Name Column -->
     <ng-container matColumnDef="fecha_salida">
       <th mat-header-cell *matHeaderCellDef mat-sort-header>Fecha de Salida</th>
       <td mat-cell *matCellDef="let row">{{row.fecha_salida}}</td>
     </ng-container>
 
-    <!-- Name Column -->
     <ng-container matColumnDef="descripcion">
       <th mat-header-cell *matHeaderCellDef mat-sort-header>Estado de Vuelo</th>
       <td mat-cell *matCellDef="let row">{{row.descripcion}}</td>
     </ng-container>
+
     <ng-container matColumnDef="Acciones">
       <th mat-header-cell *matHeaderCellDef mat-sort-header></th>
       <td mat-cell *matCellDef="let row">
         <a class="puntero" style="color :lightskyblue" matTooltip="Ver Persona">
-          <button class="btn btn-info">Seleccionar</button>
+          <button class="btn btn-info" (click)="seleccionarVuelo(row.id_vuelo,row.id_ruta)" >Seleccionar</button>
         </a>
 
       </td>
