--- src/app/componentes/vuelo/mostrar-tabla/mostrar-tabla.component.html
@@ -1,1 +1,0 @@
-<p>mostrar-tabla works!</p>
