--- src/app/componentes/vuelo/mostrar-tabla/mostrar-tabla.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-mostrar-tabla',
-  templateUrl: './mostrar-tabla.component.html',
-  styleUrls: ['./mostrar-tabla.component.css']
-})
-export class MostrarTablaComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
