--- src/app/componentes/vuelo/services/vuelo.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { VueloService } from './vuelo.service';
-
-describe('VueloService', () => {
-  let service: VueloService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(VueloService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
