--- src/app/componentes/vuelo/services/vuelo.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class VueloService {
-
-  constructor() { }
-}
