+++ src/app/nav/nav.component.html
@@ -6,8 +6,8 @@
       [opened]="(isHandset$ | async) === false">
     <mat-toolbar>Menu</mat-toolbar>
     <mat-nav-list>
-      <a mat-list-item routerLink="/vuelo">Vuelos</a>
-      <a mat-list-item routerLink="/disponible">Disponile</a>
+      <a mat-list-item routerLink="/">Inicio</a>
+      <a mat-list-item routerLink="/disponible">Disponible</a>
       <a mat-list-item routerLink="/disponibles">Link 3</a>
     </mat-nav-list>
   </mat-sidenav>
@@ -24,7 +24,6 @@
       <span class="example-spacer"></span>
       <a mat-button href="#inicio">INICIO</a>
     </mat-toolbar>
-    <!-- Add Content Here -->
     <router-outlet></router-outlet>
 
   </mat-sidenav-content>
